Add files via upload
In this homework. We have 2 tasks and for the first task, we have to calculate mathematical 
operations to define the similarity and dissimilarity between an image patch P and a template image 
T. These mathematical operations are Correlation, Zero-mean Correlation, and Sum of Squared difference, Normalized Cross-Correlation In the second task I compare the Butterworth low pass and high pass filter results difference.